test: move module action tests into own file
